+++ src/rx-query.ts
@@ -19,7 +19,8 @@ import {
     PROMISE_RESOLVE_FALSE,
     RXJS_SHARE_REPLAY_DEFAULTS,
     ensureNotFalsy,
-    areRxDocumentArraysEqual
+    areRxDocumentArraysEqual,
+    promiseWait
 } from './plugins/utils/index.ts';
 import {
     newRxError,
@@ -676,6 +677,21 @@ export async function queryCollection<RxDocType>(
     const collection = rxQuery.collection;
 
     /**
+     * If a change happens during the query-run,
+     * we do not know for 100% if that change is already included
+     * into the query results or not. The storage itself does not give that information.
+     * This lead to cases where the query results where outdated but RxDB thought
+     * that the changeevents must not be processed.
+     * To fix this we re-run the query if a change happens directly during the query run.
+     *
+     * @link https://github.com/pubkey/rxdb/issues/7067
+     */
+    let eventsDuringQueryRun = 0;
+    const sub = collection.eventBulks$.subscribe(() => {
+        eventsDuringQueryRun++;
+    });
+
+    /**
      * Optimizations shortcut.
      * If query is find-one-document-by-id,
      * then we do not have to use the slow query() method
@@ -722,6 +738,13 @@ export async function queryCollection<RxDocType>(
         const queryResult = await collection.storageInstance.query(preparedQuery);
         docs = queryResult.documents;
     }
+
+    sub.unsubscribe();
+    if (eventsDuringQueryRun > 0) {
+        await promiseWait(0);
+        return queryCollection(rxQuery);
+    }
+
     return {
         docs,
         counter: collection._changeEventBuffer.getCounter()